requirements.txt
flask
flask_jwt_extended
flask_sqlalchemy
flask-smorest
sqlalchemy
marshmallow
python-dotenv
passlib
pymysql